move jsonconvert singleton to the 'convert' directory
dedicated to the converters, so put it with them
this also hides implementation details
