+++ src/component/parser/json/converter/AbstractConverter.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 import { JsonConvert, JsonConverter, JsonCustomConvert } from 'json2typescript';
-import JsonParser from '../JsonParser';
+import JsonParser from './JsonParser';
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export function ensureIsArray(elements: Array<any> | any): Array<any> {
